Share crank cadence calculation and always show wheel circumference
