+++ Moblin/Various/Settings/Settings.swift
@@ -642,7 +642,7 @@ class SettingsWorkoutDevice: Codable, Identifiable, ObservableObject, Named {
     @Published var enabled: Bool = false
     @Published var bluetoothPeripheralName: String?
     @Published var bluetoothPeripheralId: UUID?
-    @Published var wheelCircumferenceMillimeters: Int = defaultWheelCircumferenceMillimeters
+    @Published var wheelCircumference: Int = defaultWheelCircumference
 
     enum CodingKeys: CodingKey {
         case id
@@ -650,7 +650,7 @@ class SettingsWorkoutDevice: Codable, Identifiable, ObservableObject, Named {
         case enabled
         case bluetoothPeripheralName
         case bluetoothPeripheralId
-        case wheelCircumferenceMillimeters
+        case wheelCircumference
     }
 
     func encode(to encoder: any Encoder) throws {
@@ -660,7 +660,7 @@ class SettingsWorkoutDevice: Codable, Identifiable, ObservableObject, Named {
         try container.encode(.enabled, enabled)
         try container.encode(.bluetoothPeripheralName, bluetoothPeripheralName)
         try container.encode(.bluetoothPeripheralId, bluetoothPeripheralId)
-        try container.encode(.wheelCircumferenceMillimeters, wheelCircumferenceMillimeters)
+        try container.encode(.wheelCircumference, wheelCircumference)
     }
 
     init() {}
@@ -672,9 +672,9 @@ class SettingsWorkoutDevice: Codable, Identifiable, ObservableObject, Named {
         enabled = container.decode(.enabled, Bool.self, false)
         bluetoothPeripheralName = try? container.decode(String.self, forKey: .bluetoothPeripheralName)
         bluetoothPeripheralId = try? container.decode(UUID.self, forKey: .bluetoothPeripheralId)
-        wheelCircumferenceMillimeters = container.decode(.wheelCircumferenceMillimeters,
-                                                         Int.self,
-                                                         defaultWheelCircumferenceMillimeters)
+        wheelCircumference = container.decode(.wheelCircumference,
+                                              Int.self,
+                                              defaultWheelCircumference)
     }
 }
